backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
chromadb==0.4.22
sentence-transformers==2.3.1
groq==0.4.2
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
httpx==0.26.0
tenacity==8.2.3
